// src/components/pages/EarnAPY/EarnAPY.types.ts
import { IFooterProps } from '../../atoms/Footer/Footer.types';
import { INavbarProps } from '../../atoms/Navbar/Navbar.types';

export interface IEarnAPYProps {
  eyebrow: string;
  title: string;
  caption: string;
  navbarProps: INavbarProps;
  footerProps: IFooterProps;
  cardTitle: string;
  button1Text: string;
  button2Text: string;
  button3Text: string;
  cardBody: string;
  cardValue: string;
  cardBody2: string;
  cardValue2: string;
  input1Label: string;
  input1Placeholder: string;
  input2Label: string;
  input2Placeholder: string;
}